Conflicts:
	hw/xwin/winclipboard/internal.h
	hw/xwin/winclipboard/xevents.c
	hw/xwin/winconfig.c
	hw/xwin/winkeybd.c

+++ hw/xwin/wmutil/keyboard.c
@@ -63,7 +63,7 @@ winTranslateKey(WPARAM wParam, LPARAM lParam)
     int iParamScanCode = LOBYTE(iParam);
     int iScanCode;
 
-    winDebug("winTranslateKey: wParam %08x lParam %08x\n", wParam, lParam);
+    winDebug("winTranslateKey: wParam %08x lParam %08x\n", (unsigned int)wParam, (unsigned int)lParam);
 
 /* WM_ key messages faked by Vista speech recognition (WSR) don't have a
  * scan code.
